Implement Google sign in
Allow users to sign in using their Google account and ensure that the sign in page properly redirects when it gets re-rendered. Also added .env file to prevent sensitive information from being included in version control.

Important note: this does not exclude the .env file contents from the final build and is only to be considered secure during development, as compilation will include these variables as plain text in the app code.